Add fetch_block, start_sm functions, add state machine
