+++ .aws-sam/build.toml
@@ -1,12 +1,22 @@
 # This file is auto generated by SAM CLI build command
 
-[function_build_definitions.3b57a9dd-04a0-4000-92d6-4d69bf2dd647]
-codeuri = "/Users/maxcopeland/projects/thesis/danaides/hello_efs"
+[function_build_definitions]
+[function_build_definitions.2a1a39be-41f1-448c-9359-a6c06db7581a]
+codeuri = "/Users/maxcopeland/projects/thesis/danaides/src/fetch_block"
 runtime = "python3.9"
 architecture = "x86_64"
-handler = "app.lambda_handler"
+handler = "handler.lambda_handler"
 manifest_hash = "d41d8cd98f00b204e9800998ecf8427e"
 packagetype = "Zip"
-functions = ["HelloEfsFunction"]
+functions = ["FetchBlockFunction"]
+
+[function_build_definitions.dd80837c-5516-4871-a21f-1200d43c3193]
+codeuri = "/Users/maxcopeland/projects/thesis/danaides/src/start_sm"
+runtime = "python3.9"
+architecture = "x86_64"
+handler = "handler.lambda_handler"
+manifest_hash = "d41d8cd98f00b204e9800998ecf8427e"
+packagetype = "Zip"
+functions = ["StartStateMachineFunction"]
 
 [layer_build_definitions]
