+++ template.yaml
@@ -12,11 +12,15 @@ Globals:
     Tracing: Active
   Api:
     TracingEnabled: true
+
+    
 Resources:
+
   EfsLambdaVpc:
     Type: AWS::EC2::VPC
     Properties:
       CidrBlock: 10.0.0.0/16
+
   EfsLambdaSecurityGroup:
     Type: AWS::EC2::SecurityGroup
     Properties:
@@ -32,6 +36,7 @@ Resources:
         FromPort: 0
         ToPort: 65535
         IpProtocol: tcp
+
   EfsLambdaSubnetA:
     Type: AWS::EC2::Subnet
     Properties:
@@ -39,6 +44,7 @@ Resources:
       AvailabilityZone: !Select [0, !GetAZs '']
       MapPublicIpOnLaunch: false
       CidrBlock: 10.0.0.0/24
+
   EfsLambdaSubnetB:
     Type: AWS::EC2::Subnet
     Properties:
@@ -46,8 +52,10 @@ Resources:
       AvailabilityZone: !Select [1, !GetAZs '']
       MapPublicIpOnLaunch: false
       CidrBlock: 10.0.1.0/24
+
   EfsFileSystem:
     Type: AWS::EFS::FileSystem
+
   MountTargetA:
     Type: AWS::EFS::MountTarget
     Properties:
@@ -55,6 +63,7 @@ Resources:
       SubnetId: !Ref EfsLambdaSubnetA
       SecurityGroups:
       - !Ref EfsLambdaSecurityGroup
+
   MountTargetB:
     Type: AWS::EFS::MountTarget
     Properties:
@@ -62,6 +71,7 @@ Resources:
       SubnetId: !Ref EfsLambdaSubnetB
       SecurityGroups:
       - !Ref EfsLambdaSecurityGroup
+
   AccessPoint:
     Type: AWS::EFS::AccessPoint
     Properties:
@@ -75,14 +85,43 @@ Resources:
           OwnerGid: '1000'
           OwnerUid: '1000'
           Permissions: '755'
-  HelloEfsFunction:
+
+  StateMachine:
+    Type: AWS::Serverless::StateMachine
+    Properties:
+      DefinitionUri: statemachine.asl.json
+      Policies:
+        - Version: "2012-10-17"
+          Statement:
+            - Effect: Allow
+              Action:
+                - "cloudwatch:*"
+                - "logs:*"
+              Resource: "*"
+
+  StartStateMachineFunction: 
+    Type: AWS::Serverless::Function
+    Properties:
+      CodeUri: src/start_sm/
+      Handler: handler.lambda_handler
+      Runtime: python3.9
+      Architectures:
+      - x86_64
+      Events:
+        HelloWorld:
+          Type: Api
+          Properties:
+            Path: /hello
+            Method: get
+
+  FetchBlockFunction:
     Type: AWS::Serverless::Function
     DependsOn:
     - MountTargetA
     - MountTargetB
     Properties:
-      CodeUri: hello_efs/
-      Handler: app.lambda_handler
+      CodeUri: src/fetch_block/
+      Handler: handler.lambda_handler
       Runtime: python3.9
       Architectures:
       - x86_64
@@ -99,12 +138,6 @@ Resources:
       FileSystemConfigs:
       - Arn: !GetAtt AccessPoint.Arn
         LocalMountPath: /mnt/lambda
-      Events:
-        HelloWorld:
-          Type: Api
-          Properties:
-            Path: /hello
-            Method: get
 
   ApplicationResourceGroup:
     Type: AWS::ResourceGroups::Group
